feat(cpp-campaign-v2): union-find — uint16_t parents, path halving, smaller-root-wins (W2)
Use uint16_t parent storage with constructor size validation.

Compress find paths with path halving and debug bounds assertions.

Union roots deterministically by reparenting the larger root to the smaller root.

Cover chain-order invariance, star/disjoint sets, explicit tie-breaks, repeated determinism, roots, and debug bounds.

+++ tiles-maxxing/cpp-campaign-v2/include/campaign/union_find.h
@@ -56,7 +56,7 @@ class DSU {
   std::vector<std::int32_t> roots() const;
 
  private:
-  mutable std::vector<std::int32_t> parent_;
+  mutable std::vector<std::uint16_t> parent_;
 };
 
 }  // namespace campaign
